requirements.txt
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.13.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.1
idna==3.6
openai==1.14.3
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
